rate: simplify code
When a long time has passed since a previous call to the rate limiter,
we can overflow when calculating how many new tokens are available.

To prevent that, we calculated a maximum elapsed amount.
When we hit maxElapsed, we ended up converting a duration into tokens
and then back. This operation was lossy, so we took pains to increase
the precision of the conversion in tokensFromDuration.

An easier fix is to avoid the round trip entirely.
We know the max number of new tokens we should ever have,
so we can cap it directly. And in fact we do cap it already.
Delete the extraneous code.

This lets us revert to the original, simpler tokensFromDuration.

While we're here, tweak durationFromTokens for clarity.

Updates #34861